Faster temperature sampling and moving average
There is a large noise in the temperature readings. That makes the projected temperatures unreliable.

Here we add a moving average of 100 samples on top of the temperature readings. The value of 100 was found by looking at the data from previous runs. The temperature readings are done every 10 ms. Assuming the noise is high-frequency, this should produce smooth temperature readings. The added latency is no more than 1 second, which is negligible.

The heatings controls are updated every 2 seconds. The logic is not changed.

We add an element to the serial output to log which heater has been chosen.